partial jack_transport support

+++ teq/teq.h
@@ -90,7 +90,7 @@ namespace teq
 		
 		transport_position m_transport_position;
 		
-		float m_time_until_next_tick;
+		double m_time_until_next_tick;
 		
 		
 		bool m_send_all_notes_off_on_loop;
